// src/enums/actions.enum.ts
export enum Actions {
  CAN_EVALUATE = 1,
  CAN_LOAD_FILES = 2,
  CAN_VIEW = 3,
  CAN_VIEW_COMMENTS = 4,
  CAN_VIEW_SUMMARY = 6,
  CAN_CREATE_SUMMARY = 7,
  CAN_APPROVE_CAMPUS = 10,
  CAN_SEND_TO_REVIEW = 11,
  CAN_CREATE_CURRICULUM = 12,
  CAN_VIEW_CURRICULUM = 13,
  CAN_VIEW_PROFILE_ENTRY = 14,
  CAN_CREATE_PROFILE_ENTRY = 15,
  CAN_CREATE_RAE = 17,
  CAN_VIEW_COMPETENCIES = 18,
  CAN_CREATE_COMPETENCIES = 19,
  CAN_VIEW_OBJECTIVES = 20,
  CAN_CREATE_OBJECTIVES = 21,
  CAN_VIEW_SILABUS = 22,
  CAN_CREATE_SILABUS = 23,
  CAN_VIEW_ENGLISH_FORMATION = 24,
  CAN_CREATE_ENGLISH_FORMATION = 25,
  CAN_VIEW_CAMPUS_ACADEMIC_PROGRAMS = 26,
  CAN_CREATE_CAMPUS_ACADEMIC_PROGRAMS = 27,
  CAN_CREATE_EXTENSION_SOCIAL_PROJECTION = 28,
  CAN_VIEW_EXTENSION_SOCIAL_PROJECTION = 29,
  CAN_CREATE_INTERNATIONAL = 30,
  CAN_VIEW_INTERNATIONAL = 31,
  CAN_CREATE_INVESTIGATIVE_INFORMATION = 32,
  CAN_VIEW_INVESTIGATIVE_INFORMATION = 33,
  CAN_EVALUATE_SUMMARY = 34,
}